+++ src/main/kotlin/blackjack/domain/generator/CardGenerator.kt
@@ -1,6 +1,6 @@
 package blackjack.domain.generator
 
-import blackjack.domain.model.Card
+import blackjack.domain.model.card.Card
 
 fun interface CardGenerator {
     fun create(): List<Card>
