+++ src/test/kotlin/blackjack/domain/model/CardTest.kt
@@ -1,5 +1,8 @@
 package blackjack.domain.model
 
+import blackjack.domain.model.card.Card
+import blackjack.domain.model.card.CardNumber
+import blackjack.domain.model.card.Suit
 import org.junit.jupiter.api.assertDoesNotThrow
 import kotlin.test.Test
