+++ src/test/kotlin/blackjack/domain/model/RuleTest.kt
@@ -1,5 +1,6 @@
 package blackjack.domain.model
 
+import blackjack.domain.model.card.Card
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
